Added ObjectId validation to update and delete routes

+++ server/todo_api.py
@@ -9,11 +9,13 @@ app.config["MONGODB_SETTINGS"] = {
     "db_name" : "todo_db"
 }
 
+
 @app.before_request
 def init_db():
     if "db_client" not in g:
         current_app.config["db_client"] = pymongo.MongoClient(current_app.config["MONGODB_SETTINGS"]["connection_uri"])
         g.db = current_app.config["db_client"][current_app.config["MONGODB_SETTINGS"]["db_name"]]
+
 
 @app.teardown_request
 def close_db(exception):
@@ -22,6 +24,7 @@ def close_db(exception):
 
     if client is not None:
         client.close()
+
 
 @app.route("/read_todo/<todo_id>")
 def read_todo(todo_id):
@@ -35,6 +38,7 @@ def read_todo(todo_id):
         return "Todo not found", 404
     else:
         return {"id":str(todo["_id"]), "todo":todo["todo"]}, 200
+
 
 @app.route("/read_todos")
 def read_todos():
@@ -56,8 +60,12 @@ def write_todo():
 
     return {"id":str(id.inserted_id), "todo":insert_data.get("todo")}, 200
 
+
 @app.route("/update_todo/<todo_id>", methods=["PUT"])
 def update_todo(todo_id):
+    if not ObjectId.is_valid(todo_id):
+        return "Invalid todo id", 400
+    
     new_data = request.get_json()
     todo_collecion = g.db.todos
 
@@ -69,6 +77,9 @@ def update_todo(todo_id):
 
 @app.route("/delete_todo/<todo_id>", methods=["DELETE"])
 def delete_todo(todo_id):
+    if not ObjectId.is_valid(todo_id):
+        return "Invalid todo id", 400
+    
     todo_collection = g.db.todos
 
     result = todo_collection.delete_one({"_id":ObjectId(str(todo_id))})
